Reworked MACD
Added the RSI flag back in for the buy conditions and edited the MACD conditions to be more in line with how they use it in the example, however in the buy conditions for MACD added rsiBool to see if the rsi conditions can be adjusted. Also added MathAbs(MacdCurrent) part the 3 was taken from the tutorial which was their level to avoid what they refer to as "microhills" basically to help find a stronger uptrend and remove the sideways signals.