Short Position Update 0.2
Still Working in getting the trailing stop loss to effectively get set. This will run with the trailing stoploss taking effect however the stopLoss is set so that it is 1000 * _Point away from the ask. This means some of the trades for the short positions end up selling for losses. Working to have it work the same as the Buying conditions.