Updated Short Selling
Fixed the trailing stop missing sell orders as well as kind of fixed the issue with the trailing stop needing a stoploss other than zero to have it sort of function like the buying trailing stop. However as a result kind of built a death star port into the code but that can be addressed later.